register type adapters in ApiClient class
